CA-5:  Completed development of Region/State unit tests.  Modified API and DAO to reflect bettter programming style.

+++ api_parent/api_addressbook/src/main/java/org/modules/postal/PostalApiImpl.java
@@ -346,6 +346,10 @@ class PostalApiImpl extends AbstractTransactionApiImpl implements PostalApi {
      */
     @Override
     public int deleteRegion(int stateId) throws PostalApiException {
+        if (stateId <= 0) {
+            this.msg = "State id is invalid...must be greater than zero";
+            throw new PostalApiException(this.msg);
+        }
         RegionCountryDao dao = this.factory.createRmt2OrmRegionCountryDao(this.appName);
         try {
             return dao.deleteRegion(stateId);
